requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
requests
python-multipart
